fix bug

+++ src/main/java/util/OutputFormatter.java
@@ -96,7 +96,7 @@ public class OutputFormatter {
         return String.format("\tItem number: %d, Supply warehouse ID: %d, Quantity: %d, Price: %.2f, Datetime: %s",
                 itemInfo.getInt("OL_I_ID"),
                 itemInfo.getInt("OL_SUPPLY_W_ID"),
-                itemInfo.getInt("OL_QUANTITY"),
+                itemInfo.getDecimal("OL_QUANTITY").intValue(),
                 itemInfo.getDecimal("OL_AMOUNT").doubleValue(),
                 itemInfo.getTimestamp("OL_DELIVERY_D").toString());
     }
